separate out scheduler to a module
